% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{missPLS-package}
\alias{missPLS}
\alias{missPLS-package}
\title{missPLS: Methods and Reproducible Workflows for Partial Least Squares with Missing Data}
\description{
Methods-first tooling for reproducing and extending the partial least squares regression studies on incomplete data described in Nengsih et al. (2019) \doi{10.1515/sagmb-2018-0059}. The package provides simulation helpers, missingness generators, imputation wrappers, component-selection utilities, real-data diagnostics, and reproducible study orchestration for Nonlinear Iterative Partial Least Squares (NIPALS)-Partial Least Squares (PLS) workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbertran.github.io/missPLS/}
  \item \url{https://github.com/fbertran/missPLS}
  \item Report bugs at \url{https://github.com/fbertran/missPLS/issues}
}

}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net}

Authors:
\itemize{
  \item Titin Agustin Nengsih
  \item Myriam Maumy-Bertrand
}

}
\keyword{internal}
